+++ packages/test/fixtures/parser/qualified-rule-invalid/error/index.json
@@ -1,5 +1,5 @@
 {
   "message": [
-    "Encounter unclosed block"
+    "Encounter unclosed block when consuming a rule"
   ]
 }
